CSC2033_Team04_23-24-main/requirements.txt
Flask~=3.0.0
flask_sqlalchemy
flask_login
flask_wtf
WTForms~=3.1.0
email_validator
python-dotenv~=1.0.0
pyotp~=2.9.0
flask-qrcode
bcrypt~=4.1.1
Flask-Talisman
pywebview
pillow
pymysql
pytest~=8.2.0
MarkupSafe~=2.1.3
SQLAlchemy~=2.0.21
pymongo
flask-mailman
Flask-Migrate
Flask-APScheduler
matplotlib~=3.9.0
itsdangerous~=2.2.0
alembic~=1.13.1
playwright
pytest-playwright
Flask-APScheduler
APScheduler
certifi
